% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitClustersRelatedFunctions.R
\name{align}
\alias{align}
\title{Multiple alignment with mafft.}
\usage{
align(rf, type = "AA", n_threads = 1L, accu = TRUE, mxit1000 = TRUE)
}
\arguments{
\item{rf}{A \code{list} of \code{SeqFastaAA} objects.}

\item{type}{\code{character}. One of 'AA' or 'DNA'.}

\item{n_threads}{\code{integer}. The number of threads to use.}

\item{accu}{\code{logical}. Accurate alignment?}

\item{mxit1000}{\code{logical}. Set \code{mafft} parameter
\code{--maxiterate 1000} .}
}
\value{
A \code{matrix} with columns as positions in the alignment and rows
as sequences.
}
\description{
Takes a list of \code{SeqFastaAA} objects from \link{seqinr}
package and performs a multiple alignment with mafft aligner. Outputs a
\code{matrix} with columns as positions in the alignment and rows as
sequences.
}
\references{
Katoh, Misawa, Kuma, Miyata (2002). MAFFT: a novel method for
rapid multiple sequence alignment based on fast Fourier transform. \emph{Nucleic
Acids Res} \strong{30}:3059-3066.
}
\author{
Ignacio Ferres
}
