% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realocateSingletons.R
\name{assignOrphans}
\alias{assignOrphans}
\title{Assign Orphan Clusters to other clusters.}
\usage{
assignOrphans(m, final.clusters, panm)
}
\arguments{
\item{m}{A tblout \code{data.frame}.}

\item{final.clusters}{A \code{list} of clusters.}

\item{panm}{A \code{data.frame}. The panmatrix.}
}
\value{
A \code{list} of refined clusters.
}
\description{
.
}
\author{
Ignacio Ferres
}
