% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{buildPanMatrix}
\alias{buildPanMatrix}
\title{Build the pan-matrix}
\usage{
buildPanMatrix(pangenome, type = "binary")
}
\arguments{
\item{pangenome}{A \code{list} of protein clusters or a \code{pangenome}
object as returned by \link{pangenome}.}

\item{type}{One of: "none" (no panmatrix), "binary" (presence/absence
matrix) (DEFAULT), "nparalog" (number of paralogues for each organism and
orthologous group), "representative" (just one sequence of each organism, in
case it has more than one paralogue), or "allgenes" (write all sequences,
paralogues included).}
}
\value{
A pan-genome \code{data.frame}. Each organism is a column, and
each orthologue cluster is a row. If \code{binary=T}, the values (0 or 1)
respectively. Otherwise each cell contain the number of paralogues of each
orthologue group for each organism.
}
\description{
Builds the pan-matrix from the list of orthologue clusters.
}
\details{
The function detects if \code{pangenome} argument is a list
(as used internally by \code{pangenome()} function, or returned  by
\link{getClusterGenes} function with params \code{annot=F} and
\code{sequence=F}) or a \code{pangenome} object as returned by the cited
function (the most probable case for the user).
}
\author{
Ignacio Ferres
}
