% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{checkIfParalogues}
\alias{checkIfParalogues}
\title{Check if cluster contains paralogues}
\usage{
checkIfParalogues(p)
}
\arguments{
\item{p}{A vector of proteins}
}
\value{
\code{logical}, if contains paralogues (i.e. proteins of the same
organism).
}
\description{
Check if a cluster contains paralogues.
}
\author{
Ignacio Ferres
}
