% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterOrphansRelatedFunctions.R
\name{clusterOrphans}
\alias{clusterOrphans}
\title{Cluster 'orphan' sequences (i.e. without Pfam-A domains of classes
'Domain' or 'Family' found).}
\usage{
clusterOrphans(tout, fastas, n_threads)
}
\arguments{
\item{tout}{A \code{data.frame} with non overlapping domain hmmscan output
as passed by \link{processHmmsearch}.}

\item{fastas}{A \code{list} of protein sequeces as \code{SeqFastaAA} objects}

\item{n_threads}{\code{integer} Number of threads to use.}
}
\value{
A \code{list} of protein clusters.
}
\description{
In order to avoid exponencial number of comparisons with the
number of sequences, an heuritic is implemented. It first takes all the
proteins of the 3 organisms which have more orphans and performs phmmer
comparisons against all orphans. The phmmer output is saved. Then the Markov
Clustering algorithm (MCL) clusters the sequences and those groups which
contain proteins of the first 3 organisms are provisionally removed from the
analysis. Then another round of phmmer comparisons is performed with the
proteins of the 3 organisms with more orphans left against all orphans left,
saves the phmmer output, and MCL clusters the sequences.

The same is repeated until there are no more sequences left. Then it merge
all the saved phmmer output tables and performs the final MCL clustering
over it.
}
\author{
Ignacio Ferres
}
