% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{coreAlign}
\alias{coreAlign}
\title{Align Core-Genes}
\usage{
coreAlign(x, ffns = c(), level = 1, accu = TRUE, n_threads = 1L,
  file.out = "")
}
\arguments{
\item{x}{A \code{pangenome} object.}

\item{ffns}{A vector of nucleotide fasta files.}

\item{level}{\code{numeric} between 1 and 0.9.}

\item{accu}{\code{logical}. Accurate alignment?}

\item{n_threads}{\code{integer}. The number of threads to use.}

\item{file.out}{\code{character}. The name of the output file.}
}
\value{
A fasta file with the concatenated core-genes alignment.
}
\description{
Align core-genes and outputs a concatenation of the aligned
genes. Core-genes are specified at a certain level of permissiveness.
}
\details{
This function performs an alignment of each orthologous group and
concatenates them into a "super-gene". This is specially suitable to perform
subsecuent phylogenetic analysis.
}
\author{
Ignacio Ferres
}
