% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmsearchRelatedFunctions.R
\name{determineOverlap}
\alias{determineOverlap}
\title{Determine which Pfam domains overlaps.}
\usage{
determineOverlap(m)
}
\arguments{
\item{m}{A logical \code{matrix} indicating if 'x' domain overlaps with 'y'.}
}
\value{
A \code{list} of overlapping domains.
}
\description{
Takes an overlappling-domain logical matrix \code{m} and
through a 'trace-forward' precedure from \code{m[1,1]} to \code{m[ ,ncol(m)]}
outputs a list with overlappling-domains row names.
}
\author{
Ignacio Ferres
}
