% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{domainSearch}
\alias{domainSearch}
\title{Compute All Pfam Related Functions}
\usage{
domainSearch(fastas, hmmPfam, datPfam, n_threads = 1L)
}
\arguments{
\item{fastas}{A \code{list} of amino acid sequences.}

\item{hmmPfam}{\code{character} The path to the hmm file.}

\item{datPfam}{\code{character} The path to the dat file.}

\item{n_threads}{\code{integer} The number of cpus to use.}
}
\value{
a \code{list} with the domain and family clustering.
}
\description{
Computes all Pfam-A related functions.
}
\author{
Ignacio Ferres
}
