% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getClusterGenes}
\alias{getClusterGenes}
\title{Extract the genes of a set of clusters}
\usage{
getClusterGenes(x, clustNames = c(), annot = F)
}
\arguments{
\item{x}{A \code{pangenome} object.}

\item{clustNames}{A \code{vector} of cluster names.}

\item{annot}{\code{logical} Should original gene annotation be retrieved?
of \link{seqinr} package be retrieved?}
}
\value{
A \code{list} of ethier internal gene names, original annotation
names, or the sequences as class \code{SeqFastaAA} from the clusters names
passed in \code{clustNames} argument.
}
\description{
Extract a list of genes from a set of clusters in the
\code{pangenome} object.
}
\details{
Something
}
\author{
Ignacio Ferres
}
