% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{getCoreClusters}
\alias{getCoreClusters}
\title{Extract core cluster names}
\usage{
getCoreClusters(x, level = 1L)
}
\arguments{
\item{x}{A \code{pangenome} object.}

\item{level}{\code{numeric} A number between 0.9 and 1.}
}
\value{
A \code{vector} of the core genes cluster names at the specified
level permissiveness.
}
\description{
Extract core cluster names at a certain level of permissiveness
}
\details{
Extracts the core cluster names at a certain level of
permissiveness.
\code{level} from 1 to 0.99 is the strict definition of core-genome, which
is the set of genes shared by the 100%-99% of the organisms.

A \code{level}=0.95 extracts the softcore clusters, i.e. those clusters that
contain genes shared by a 95% of the organisms.

This level can be adjusted to up to 0.9, which is only recomended in case of
working with very distant species or in case of bad sequence qualities so a
relatively big proportion of genes (proteins) may be absent or truncated in
the original genome sequence.
}
\author{
Ignacio Ferres
}
