% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmsearch_related.R
\name{grouping}
\alias{grouping}
\title{Split hmmscan output by query.}
\usage{
grouping(t, singlehit = FALSE, type = "Domain")
}
\arguments{
\item{t}{\code{data.frame}. hmmscan output passed by \link{outhmmsearch}.}

\item{singlehit}{\code{logical}. Has this query just one hit?}

\item{type}{\code{character}. One of 'Domain', 'Family', 'Motif', 'Repeat',
'Disordered' or 'Coiled-coil'.}
}
\value{
A \code{list} of \code{data.frame} with hits by query.
}
\description{
Split hmmscan output by query.
}
\author{
Ignacio Ferres
}
