% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realocateSingletons.R
\name{hmmBuild}
\alias{hmmBuild}
\title{hmmBuild}
\usage{
hmmBuild(ali, name)
}
\arguments{
\item{ali}{\code{character} A fasta file name.}

\item{name}{\code{character} The name of the model.}
}
\value{
The name of the hmm model file.
}
\description{
Wrapper function of \code{hmmbuild} (HMMER 3).
}
\author{
Ignacio Ferres
}
