% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitClustersRelatedFunctions.R
\name{maxTrueOGs}
\alias{maxTrueOGs}
\title{Maximum subtrees which contains only true orthologues}
\usage{
maxTrueOGs(tru.or, l)
}
\arguments{
\item{tru.or}{A \code{vector} of logical values.}

\item{l}{A \code{list} of trees of class \code{phylo}.}
}
\value{
A logical \code{vector} indicating if true orthologue subtree is
maximum or not.
}
\description{
Detect maximum subtrees which contains only true orthologues.
}
\author{
Ignacio Ferres
}
