% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{outhmmscan}
\alias{outhmmscan}
\title{Process hmmscan output to make it 'R'eadable.}
\usage{
outhmmscan(pouti, ref)
}
\arguments{
\item{pouti}{\code{character}. hmmscan output temporary file.}

\item{ref}{\code{data.frame}. Information about each Pfam-A entry.}
}
\value{
A \code{data.frame} with the hmmscan output plus information about
Pfam-A hits.
}
\description{
Process hmmscan output to make it readable by R.
}
\note{
Taken and adapted from \code{micropan} package (Lars Snipen and
Kristian Hovde Liland).
}
\author{
Ignacio Ferres
}
