% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmsearch_related.R
\name{outhmmsearch}
\alias{outhmmsearch}
\title{Process hmmsearch output to make it 'R'eadable.}
\usage{
outhmmsearch(pouti, ref)
}
\arguments{
\item{pouti}{\code{character}. hmmsearch output temporary file.}

\item{ref}{\code{data.frame}. Information about each Pfam-A entry.}
}
\value{
A \code{data.frame} with the hmmsearch output plus information about
Pfam-A hits.
}
\description{
Process hmmsearch output to make it readable by R.
}
\note{
Taken and adapted from \code{micropan} package (Lars Snipen and
Kristian Hovde Liland).
}
\author{
Ignacio Ferres
}
