% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orphans.R
\name{outphmmer}
\alias{outphmmer}
\title{Process phmmer output to make ir 'R'eadable.}
\usage{
outphmmer(pouti)
}
\arguments{
\item{pouti}{\code{character}. phmmer output temporary file.}
}
\value{
A \code{data.frame} with the phmmer output.
}
\description{
Process phmmer output to make it readable by R. Taken and
adapted from \code{micropan} package (Lars Snipen and Kristian Hovde
Liland).
}
\author{
Ignacio Ferres
}
