% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitClustersRelatedFunctions.R
\name{pdist.aa}
\alias{pdist.aa}
\title{Calculate p distance from a multiple alignment}
\usage{
pdist.aa(x, scaled = TRUE)
}
\arguments{
\item{x}{\code{matrix}. A multiple alignment as passed by
\link{align}.}

\item{scaled}{\code{logical}. Return scaled by alignment length?}
}
\value{
A \code{dist} object.
}
\description{
Calculate p distance from a multiple alignment.
}
\note{
Taken and adapted from \code{\link[ape]{dist.aa}}, from ape package by
Paradis et al.
}
\author{
Ignacio Ferres
}
