% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{PewitR6}
\alias{PewitR6}
\title{Pewit R6 Class}
\description{
Pewit R6 Class

Pewit R6 Class
}
\section{Super classes}{
\code{\link[pagoo:PgR6]{pagoo::PgR6}} -> \code{\link[pagoo:PgR6M]{pagoo::PgR6M}} -> \code{\link[pagoo:PgR6MS]{pagoo::PgR6MS}} -> \code{PewitR6}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PewitR6-new}{\code{PewitR6$new()}}
\item \href{#method-PewitR6-gene_context}{\code{PewitR6$gene_context()}}
\item \href{#method-PewitR6-gg_gene_context}{\code{PewitR6$gg_gene_context()}}
\item \href{#method-PewitR6-clone}{\code{PewitR6$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="add_metadata"><a href='../../pagoo/html/PgR6.html#method-PgR6-add_metadata'><code>pagoo::PgR6$add_metadata()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="drop"><a href='../../pagoo/html/PgR6.html#method-PgR6-drop'><code>pagoo::PgR6$drop()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="recover"><a href='../../pagoo/html/PgR6.html#method-PgR6-recover'><code>pagoo::PgR6$recover()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="save_pangenomeRDS"><a href='../../pagoo/html/PgR6.html#method-PgR6-save_pangenomeRDS'><code>pagoo::PgR6$save_pangenomeRDS()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="write_pangenome"><a href='../../pagoo/html/PgR6.html#method-PgR6-write_pangenome'><code>pagoo::PgR6$write_pangenome()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="cg_exp_decay_fit"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-cg_exp_decay_fit'><code>pagoo::PgR6M$cg_exp_decay_fit()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="dist"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-dist'><code>pagoo::PgR6M$dist()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_barplot"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-gg_barplot'><code>pagoo::PgR6M$gg_barplot()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_binmap"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-gg_binmap'><code>pagoo::PgR6M$gg_binmap()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_curves"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-gg_curves'><code>pagoo::PgR6M$gg_curves()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_dist"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-gg_dist'><code>pagoo::PgR6M$gg_dist()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_pca"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-gg_pca'><code>pagoo::PgR6M$gg_pca()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_pie"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-gg_pie'><code>pagoo::PgR6M$gg_pie()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="pan_pca"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-pan_pca'><code>pagoo::PgR6M$pan_pca()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="pg_power_law_fit"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-pg_power_law_fit'><code>pagoo::PgR6M$pg_power_law_fit()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="rarefact"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-rarefact'><code>pagoo::PgR6M$rarefact()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="runShinyApp"><a href='../../pagoo/html/PgR6M.html#method-PgR6M-runShinyApp'><code>pagoo::PgR6M$runShinyApp()</code></a></li>
<li><span class="pkg-link" data-pkg="pagoo" data-topic="PgR6MS" data-id="core_seqs_4_phylo"><a href='../../pagoo/html/PgR6MS.html#method-PgR6MS-core_seqs_4_phylo'><code>pagoo::PgR6MS$core_seqs_4_phylo()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PewitR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-PewitR6-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PewitR6$new(
  data,
  org_meta,
  cluster_meta,
  core_level = 95,
  sep = "__",
  sequences,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PewitR6-gene_context"></a>}}
\if{latex}{\out{\hypertarget{method-PewitR6-gene_context}{}}}
\subsection{Method \code{gene_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PewitR6$gene_context(cluster, upstream = 2, downstream = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PewitR6-gg_gene_context"></a>}}
\if{latex}{\out{\hypertarget{method-PewitR6-gg_gene_context}{}}}
\subsection{Method \code{gg_gene_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PewitR6$gg_gene_context(
  cluster,
  upstream = 2,
  downstream = 2,
  fill = "cluster",
  orient = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PewitR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PewitR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PewitR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
