% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome.R
\name{plot.pangenome}
\alias{plot.pangenome}
\title{Plot a \code{pangenome} object.}
\usage{
\method{plot}{pangenome}(x, horiz = T, border = T, las = 2, xlim = c(0,
  pretty(max(colSums(x$panmatrix)))[2]), legend.text = T,
  args.legend = list(x = "topright", bg = NULL, cex = 0.5), mar = c(4, 7, 3,
  2), plot = T, ...)
}
\arguments{
\item{x}{\code{pangenome} object.}

\item{horiz}{\code{logical}. Should barplot be drawn horizontal?}

\item{border}{\code{logical}. Should borders be drawn?}

\item{las}{\code{numeric}. see \code{\link[graphics]{par}}.}

\item{xlim}{\code{vector}. limits for the x axis.}

\item{legend.text}{\code{logical}. see \code{\link[graphics]{barplot}}.}

\item{args.legend}{\code{list} of additional arguments to pass to
\code{\link[graphics]{legend}}.}

\item{mar}{\code{vector}. see \code{\link[graphics]{par}}.}

\item{plot}{\code{logical} If FALSE, nothing is plotted.}

\item{...}{arguments to be passed to \code{\link[graphics]{barplot}}.}
}
\value{
a \code{martrix} with the count of softcore, accessory and singleton
genes for each organism used to draw the barplot.
}
\description{
Generic function to plot a \code{pangenome} object.
}
\details{
A stacked barplot is generated. Each bar corresponds to each
organisms, and the bars show the proportion of soft-core, accessory and
singletons.
}
