% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plotRarefaction}
\alias{plotRarefaction}
\title{Plot pangenome rarefaction curves}
\usage{
plotRarefaction(x, nsamp = 10)
}
\arguments{
\item{x}{A \code{pangenome} object.}

\item{nsamp}{\code{integer} The number of genomes to sample on each stage.}
}
\value{
A \code{list} of 2 \code{nsamp}*# of organisms \code{matrix} is
returned.
}
\description{
Plot core-genome and pan-genome rarefaction curves.
}
\details{
Both the number of shared genes and the total number of genes as
a function of the number of organisms sequencially added are plotted. For
each new genome added, a sample of \code{nsamp} (\code{default} 10) genomes
are evaluated with no replace.

A boxplot is drawn.
Future versions will allow more customization.

Also 2 matrices are invisibly returned, the first for the core rarefaction
curve, and the second for the pangenome rarefaction curve. Each cell is the
count of either core or pan genes for the ith sample (rows) of jth organism
(columns) added.
}
\author{
Ignacio Ferres
}
