% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome.R
\name{print.pangenome}
\alias{print.pangenome}
\title{Print Generic for Pangenome Objects}
\usage{
\method{print}{pangenome}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pangenome}.}

\item{...}{Unused.}
}
\value{
\code{character}. A compact summary of a pangenome object.
}
\description{
Generic function to print "pangenome" class objects.
}
\author{
Ignacio Ferres
}
