% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{processHmmscan}
\alias{processHmmscan}
\title{Main hmmscan output processing function}
\usage{
processHmmscan(pout, ref)
}
\arguments{
\item{pout}{\code{character}. Temporary file name (hmmscan output).}

\item{ref}{\code{data.frame}. Information about each Pfam-A entry.}
}
\value{
A \code{data.frame} with the domain structure of each sequence.
}
\description{
Outputs a data.frame with the domain structure of each sequence
(rows). It has 6 columns, each one with the semi-colon separated found
sequences of Pfam (PF) entries of the following 6 categories: 'Domain',
'Family', 'Motif', 'Repeat','Disordered' or 'Coiled-coil'.
}
\author{
Ignacio Ferres
}
