% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realocateSingletons.R
\name{readTblout}
\alias{readTblout}
\title{Read tblout Formated Files}
\usage{
readTblout(tblout)
}
\arguments{
\item{tblout}{The file name.}
}
\value{
A \code{data.frame}.
}
\description{
Read hmmsearch output.
}
\author{
Ignacio Ferres
}
