% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realocateSingletons.R
\name{realocateSingletons}
\alias{realocateSingletons}
\title{Refining Singletons}
\usage{
realocateSingletons(final.clusters, panm, fastas, n_threads, seed = numeric())
}
\arguments{
\item{final.clusters}{A \code{list} of clusters.}

\item{panm}{A binary panmatrix.}

\item{fastas}{A \code{list} of sequences.}

\item{n_threads}{\code{integer} The number of cpus to use.}

\item{seed}{\code{integer}. The algorithm uses a random step. This parameter
is used to allow reproducible results.}
}
\value{
A new \code{list} of clusters.
}
\description{
Performs the final refinement which consist in evaluate the
singletons against the accessory genes. This step is performed because it
has benn seen that previous steps tends to generate more singletons than
there really are, specially if working with many genomes (100+).
}
\author{
Ignacio Feres
}
