% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rm_overlaping_clans}
\alias{rm_overlaping_clans}
\title{Removes same-clan overlapping domains}
\usage{
rm_overlaping_clans(tping, type)
}
\arguments{
\item{tping}{A \code{data.frame} with all hits for each sequence.}

\item{type}{\code{character}. One of 'Domain', 'Family', 'Motif', 'Repeat',
'Disordered' or 'Coiled-coil'.}
}
\value{
a \code{data.frame} of non-overlapping domains.
}
\description{
Removes those overlapping domains (which belong to the same
clan) with higher e-value. In a case like (a[b)c] with
\code{which.min(evalue)=='a'}, the algorithm will keep only domain 'a', even
if 'a' doesn't overlaps with 'c'. Will also remove overlapping domains with
no clan assigned, following the same above criteria as if all no clan
assigned domains were from the same clan.
}
\author{
Ignacio Ferres
}
