% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitClustersRelatedFunctions.R
\name{rmvRecentParalogues}
\alias{rmvRecentParalogues}
\title{Remove recent paralogues}
\usage{
rmvRecentParalogues(njtree, w, d)
}
\arguments{
\item{njtree}{A tree of class \code{phylo}.}

\item{w}{A \code{vector} of nodes which contain recent paralogues, as passed
by \link{whichNodesHaveRecentParalogues}.}

\item{d}{A \code{dist} object with the pairwise p-distance between proteins
as passed by \link{pdist.aa}.}
}
\value{
A \code{list} with two elements: A tree of class \code{phylo} with
paralogues removed, and a \code{list} of paralogues removed.
}
\description{
If recent paralogues are detected, just the one with the closer
distance to the tips which are descendants of the immediate previous node,
is kept.
}
\author{
Ignacio Ferres
}
