% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmsearchRelatedFunctions.R
\name{runHmmsearch}
\alias{runHmmsearch}
\title{Run Hmmsearch (HMMER 3)}
\usage{
runHmmsearch(fasta, pfam, n_threads = 1L)
}
\arguments{
\item{fasta}{A protein fasta file.}

\item{pfam}{A hmm file. Must be pressed (see hmmpress from HMMER manual).}

\item{n_threads}{An \code{integer}. The number of cores to use.}
}
\value{
The path to a temporary file where the hmmsearch output is placed.
}
\description{
Takes a fasta file and a Hidden Markov Model profile and
performs a search of the former over the latter.
}
\author{
Ignacio Ferres
}
