% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orphans.R
\name{runMCL}
\alias{runMCL}
\title{Run MCL}
\usage{
runMCL(abc, neg.log10 = TRUE, infl = 6)
}
\arguments{
\item{abc}{\code{character}. Temporary files of abc formatted phmmer
comparisons.}

\item{neg.log10}{\code{logical}. Should the negative of the log10 value be
used as similarity measure?}

\item{infl}{\code{integer}. Inflacion value.}
}
\value{
MCL output.
}
\description{
Cluster protein sequences by running the Markov Clustering
algorithm (MCL) over the phmmer comparison scores.
}
\author{
Ignacio Ferres
}
