% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{runOnExit}
\alias{runOnExit}
\title{Removes everything if an error occur}
\usage{
runOnExit(outdir)
}
\arguments{
\item{outdir}{The normalized path of the output directory}
}
\value{
Warning message if exits before finnishing all the process.
}
\description{
Removes the output folder if an error occurs after finnishing.
}
\author{
Ignacio Ferres
}
