% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{setClusterNames}
\alias{setClusterNames}
\title{Set cluster names}
\usage{
setClusterNames(final.clusters)
}
\arguments{
\item{final.clusters}{A \code{list} of protein clusters.}
}
\value{
A \code{vector} of names for the clusters
}
\description{
Creates othologue cluster names.
}
\author{
Ignacio Ferres
}
