% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{splitAndWriteFastas}
\alias{splitAndWriteFastas}
\title{Split sequences to distribute among threads}
\usage{
splitAndWriteFastas(fastas, n_threads)
}
\arguments{
\item{fastas}{A \code{list} of aminoacid sequences in \code{SeqFastaAA}
class.}

\item{n_threads}{\code{integer} The number of threads to use.}
}
\value{
A \code{vector} of temporary file names.
}
\description{
Split sequences to distribute among threads.
}
\details{
Distributes all proteins in many files as n_threads
set in order to optimize the computing power in the following step(hmmscan).
}
\author{
Ignacio Ferres
}
