% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitClustersRelatedFunctions.R
\name{splitClusters}
\alias{splitClusters}
\title{Split clusters}
\usage{
splitClusters(clstr, accuAli = FALSE)
}
\arguments{
\item{clstr}{A \code{list} of \code{SeqFastaAA} objects.}

\item{accuAli}{\code{logical}. Perform accurate alignment? If number of
sequences is grater than 400, it is set to FALSE.}
}
\value{
A \code{list} of true orthologues protein clusters.
}
\description{
Takes coarse protein groups clustered by ethier domain
structure or by phmmer comparison plus MCL, and split them in true
orthologues groups.
}
\details{
A multiple alignment is performed by mafft aligner. The alignment
is trimmed and a p-distance matrix is calculated. A neighbour joining (NJ)
tree is inferred, and midpoint rooted.

Recent paralogues are detected by identifying those nodes which only contain
tips from the same organism, and just the one with the smaller distance to
its neighbours is kept.

Maximum subtrees which contain at most one tip of each organism are
considered true orthologues.
}
\author{
Ignacio Ferres
}
