% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitClustersRelatedFunctions.R
\name{splitPreClusters}
\alias{splitPreClusters}
\title{Split Pre Clusters}
\usage{
splitPreClusters(pre.clusters, accuAli = FALSE, n_threads = 1L)
}
\arguments{
\item{pre.clusters}{A \code{list} of clusters.}

\item{accuAli}{\code{logical}. Use accurate alignment?}

\item{n_threads}{\code{integer} The number of cpus to use.}
}
\value{
A curated \code{list} of clusters.
}
\description{
Takes coarse clusters and split them using a tree prunning
algorithm.
}
\author{
Ignacio Ferres
}
