% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome.R
\name{summary.pangenome}
\alias{summary.pangenome}
\title{Summarizing pangenome object}
\usage{
\method{summary}{pangenome}(object, ...)
}
\arguments{
\item{object}{An object of class \code{pangenome}}

\item{...}{Unused}
}
\value{
A \code{table} with: Number of organisms, Number of genes, Number of
clusters, Number of core-genes present at 100% of organisms, Number of soft-
core genes present in at least 95% of organisms, Number of accessory genes
present at less than 95% of organisms but that there are not singletones,
and Number of singletones.
}
\description{
Summary method for class \code{pangenome}
}
\author{
Ignacio Ferres
}
