% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitClustersRelatedFunctions.R
\name{whichNodeHaveRecentParalogues}
\alias{whichNodeHaveRecentParalogues}
\alias{whichNodesHaveRecentParalogues}
\title{Which nodes have recent paralogues}
\usage{
whichNodesHaveRecentParalogues(subtrees)
}
\arguments{
\item{subtrees}{A \code{list} of trees of class "phylo".}
}
\value{
A \code{vector} of nodes that contain recen paralogues.
}
\description{
Detect which nodes contain only tips from the same organism.
}
\author{
Ignacio Ferres
}
