% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{whichSubtreesAreTrueOrthologues}
\alias{whichSubtreesAreTrueOrthologues}
\title{Which subtrees are true orthologues}
\usage{
whichSubtreesAreTrueOrthologues(subtrees)
}
\arguments{
\item{subtrees}{A \code{list} of trees of class \code{phylo}.}
}
\value{
A \code{vector} of \code{logical} values indicating if subtrees
contain at most one tip of each organism.
}
\description{
Which subtrees are true orthologues, i.e. which nodes contain
at most only one tip of each organism.
}
\author{
Ignacio Ferres
}
