% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{writeClusters}
\alias{writeClusters}
\title{Write Clusters}
\usage{
writeClusters(outdir, final.clusters, filename)
}
\arguments{
\item{outdir}{Output directory}

\item{final.clusters}{A list of clusters}

\item{filename}{File name.}
}
\value{
A file 'clusters.txt' in \code{outdir}.
}
\description{
Write clusters in a fasta-like format.
}
\author{
Ignacio Ferres
}
