% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{writeFastaClusters}
\alias{writeFastaClusters}
\title{Write Cluster Sequences in Fasta Format}
\usage{
writeFastaClusters(x, clustNames = c(), ffns = c(), outdir = "",
  type = "default", n_threads = 1)
}
\arguments{
\item{x}{A \code{pangenome} object.}

\item{clustNames}{The names of the clusters (OGXXXX...) to be written.}

\item{ffns}{The fasta files (as a vector of paths, as returned by
\code{list.files} with options \code{pattern = 'ffn$'} and
\code{full.names = TRUE}).}

\item{outdir}{Where to output the files.}

\item{type}{\code{character}. One of "\code{default}" (one sequence per
genome on each fasta file), "\code{all}" (including paralogues), or
"\code{representative}" (one sequence per fasta file, a representative of
the cluster).}

\item{n_threads}{The number of cpus to use. Used when reading fasta files.}
}
\value{
Fasta formated files.
}
\description{
Write cluster's cds in fasta format.
}
\author{
Ignacio Ferres
}
