% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{writeParalogues}
\alias{writeParalogues}
\title{Write Paralogues}
\usage{
writeParalogues(outdir, final.clusters)
}
\arguments{
\item{outdir}{Output directory}

\item{final.clusters}{A list of clusters}
}
\value{
A file 'paralogues.txt' in \code{outdir}.
}
\description{
Write clusters which contain paralogues in a fasta-like format.
Only clusters which do contain paralogues are written; the genes that
were already written in 'clusters.txt' (representatives) are not re-written.
}
\author{
Ignacio Ferres
}
