% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{checkIfParalogues}
\alias{checkIfParalogues}
\title{Check if cluster contains paralogues}
\usage{
checkIfParalogues(p)
}
\arguments{
\item{p}{A vector of proteins}
}
\value{
\code{logical}, if contains paralogues (i.e. proteins of the same
organism).
}
\description{
Check if a cluster contains paralogues.
}
\details{
# Deprecated
#@name outhmmscan
#@title Process hmmscan output to make it 'R'eadable.
#@description Process hmmscan output to make it readable by R.
#@param pouti \code{character}. hmmscan output temporary file.
#@param ref \code{data.frame}. Information about each Pfam-A entry.
#@return A \code{data.frame} with the hmmscan output plus information about
Pfam-A hits.
#@note Taken and adapted from \code{micropan} package (Lars Snipen and
Kristian Hovde Liland).
#@author Ignacio Ferres
}
\author{
Ignacio Ferres
}
