% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{domainSearch}
\alias{domainSearch}
\title{Compute All Pfam Related Functions}
\usage{
domainSearch(faas, hmm_pfam, dat_pfam, n_threads = 1L, verbose = TRUE)
}
\arguments{
\item{faas}{A \code{list} of amino acid sequences.}

\item{n_threads}{\code{integer} The number of cpus to use.}

\item{hmmPfam}{\code{character} The path to the hmm file.}

\item{datPfam}{\code{character} The path to the dat file.}
}
\value{
a \code{list} with the domain and family clustering.
}
\description{
Computes all Pfam-A related functions.
}
\author{
Ignacio Ferres
}
