% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff3.R
\name{extractSeqsFromGff3}
\alias{extractSeqsFromGff3}
\title{Extract Sequences from Gff3 file}
\usage{
extractSeqsFromGff3(infile)
}
\arguments{
\item{infile}{\code{character}. The gff3 filename.}
}
\value{
A \code{DNAStringSet}.
}
\description{
Extract Sequences from Gff3 file and outputs a list of element,
one per each CDS. Each element contain the DNA gene sequence and the AA
protein sequence.
}
\author{
Gregorio Iraola and Ignacio Ferres.
}
