% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome.R
\name{pangenome}
\alias{pangenome}
\title{Compute the pangenome of a set of related organisms.}
\usage{
pangenome(gffs = c(), hmmPfam = character(), datPfam = character(),
  n_threads = 1L, dir_out = "out", writeFfns = FALSE,
  writeFastas = FALSE, pmOutfileType = "representative", alignCore = TRUE,
  accuAli = FALSE, coreLevel = 1)
}
\arguments{
\item{gffs}{A \code{vector} of gff3 files as retrieved by prokka (Seeman,
2014).}

\item{hmmPfam}{\code{character} with the path to Pfam-A.hmm file.}

\item{datPfam}{\code{character} with the path to Pfam-A.hmm.dat file.}

\item{n_threads}{\code{integer}. The number of threads to use.}

\item{dir_out}{Name of output directory to be created.}

\item{writeFfns}{\code{logical}. If should write ffn sequences. If \code{alignCore}
is set to \code{TRUE} (see below), this option is set to TRUE.}

\item{writeFastas}{\code{logical}. Write fasta files with gene sequences for
each cluster?}

\item{pmOutfileType}{The type of pan-matrix you want to be written in
\code{out.dir} folder. One of: "none" (no panmatrix), "binary" (presence/
absence matrix) (DEFAULT), "nparalog" (number of paralogues for each organism
and orthologous group), "representative" (just one sequence of each organism,
in case it has more than one paralogue), or "allgenes" (write all sequences,
paralogues included).}

\item{alignCore}{\code{logical}. Align core genes and concatenate them into
a "super-gene" alignment? Suitable for phylogenetic analysis.}

\item{accuAli}{\code{logical}. Perform an accurate alignment? If \code{TRUE}
the alignment stage could take a while depending both on the number of
genomes and on the core size.}

\item{coreLevel}{\code{numeric}. A number between 1-0.9 which determines at
what proportion of presence should clusters be considered as part of the
core-genome.}
}
\value{
An object of class \code{pangenome} which basically consist on a
\code{list} with the first element being the list of clusters, the second
element the pan-matrix as a binary \code{data.frame} with columns being the
organisms and rows the cluster names, and a series of attributes.
}
\description{
Computes the pangenome of a set of related organisms. Takes a
list of gff3 files, a path to Pfam HMM models and data associated to those
models.

It first uses hmmsearch (HMMER 3) to look for Pfam domains in the aminoacid
sequences and, where it finds them, creates a domain profile. Then clusterize
sequences with the same domain structure in protein families.

Those proteins without any domain identified are compared with phmmer
(HMMER 3) and clustered with mcl into protein families.

Protein families obtained from both two previous steps are then splited into
true orthologues by a gene tree-prunning algorithm, creating a pangenome
profile of the group of organisms.
}
\details{
A scan against Pfam-A database is performed and only those hits of
PF class 'domain' or 'family' are considered for further analysis.
If two or more domains of the same clan overlap, then the one with the
smaller e-value is kept. For those proteins which have Pfam hits, a profile
of each protein is determined by the consecutive sequence of Pfam domains,
and those proteins with the same domain structure are then clustered
together to form protein families.

Phmmer is then used to compare all those proteins with no Pfam domains
identified and the Markov Clustering algorithm (mcl) clusters them into
coarse groups (i.e. protein families).

As this package is designed to handle groups of organisms related at a
greater level than species (although it works perfectly fine with close
related clades), a paralogue split based on conserved neighbourhood was not
considered because this relationships get lose at distantly related
organisms. Instead of that it relies on the concept of orthology itself to
diferenciate between paralogues and true orthologues, this is, a gene tree
based inference.

With each protein familiy groups which contain paralogues, an alignment
followed by all-vs-all distance calculation is performed. A
neighbour-joining tree is then generated and midpoint rooted.
If recent paralogues are detected (this is, nodes which all its descendants
belong to the same organism), then all but one tips are removed from the
gene-tree and saved as recent paralogues. Just the one sequence with the
closer distance to its neighbours is kept.
Finally, the gene-tree is splited in as many sub-trees as necessary to have
the minimum set of sub-trees with just one gene of each organism.

In that way true-orthologue clusters are retrieved and saved in an object of
class "pangenome".
}
\note{
External dependencies are HMMER3, MCL and MAFFT, without them running
the function will stop with an error.
}
\author{
Ignacio Ferres and Gregorio Iraola
}
