% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.R
\name{PewitR6}
\alias{PewitR6}
\title{Pewit R6 Class}
\description{
Pewit R6 Class

Pewit R6 Class
}
\section{Super classes}{
\code{\link[pagoo:PgR6]{pagoo::PgR6}} -> \code{\link[pagoo:PgR6M]{pagoo::PgR6M}} -> \code{\link[pagoo:PgR6MS]{pagoo::PgR6MS}} -> \code{PewitR6}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PewitR6$new()}}
\item \href{#method-gene_context}{\code{PewitR6$gene_context()}}
\item \href{#method-gg_gene_context}{\code{PewitR6$gg_gene_context()}}
\item \href{#method-clone}{\code{PewitR6$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="add_metadata">}\href{../../pagoo/html/PgR6.html#method-add_metadata}{\code{pagoo::PgR6$add_metadata()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="drop">}\href{../../pagoo/html/PgR6.html#method-drop}{\code{pagoo::PgR6$drop()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="recover">}\href{../../pagoo/html/PgR6.html#method-recover}{\code{pagoo::PgR6$recover()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="save_pangenomeRDS">}\href{../../pagoo/html/PgR6.html#method-save_pangenomeRDS}{\code{pagoo::PgR6$save_pangenomeRDS()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6" data-id="write_pangenome">}\href{../../pagoo/html/PgR6.html#method-write_pangenome}{\code{pagoo::PgR6$write_pangenome()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="binomix_estimate">}\href{../../pagoo/html/PgR6M.html#method-binomix_estimate}{\code{pagoo::PgR6M$binomix_estimate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="cg_exp_decay_fit">}\href{../../pagoo/html/PgR6M.html#method-cg_exp_decay_fit}{\code{pagoo::PgR6M$cg_exp_decay_fit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="dist">}\href{../../pagoo/html/PgR6M.html#method-dist}{\code{pagoo::PgR6M$dist()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="fluidity">}\href{../../pagoo/html/PgR6M.html#method-fluidity}{\code{pagoo::PgR6M$fluidity()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_barplot">}\href{../../pagoo/html/PgR6M.html#method-gg_barplot}{\code{pagoo::PgR6M$gg_barplot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_binmap">}\href{../../pagoo/html/PgR6M.html#method-gg_binmap}{\code{pagoo::PgR6M$gg_binmap()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_curves">}\href{../../pagoo/html/PgR6M.html#method-gg_curves}{\code{pagoo::PgR6M$gg_curves()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_dist">}\href{../../pagoo/html/PgR6M.html#method-gg_dist}{\code{pagoo::PgR6M$gg_dist()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_pca">}\href{../../pagoo/html/PgR6M.html#method-gg_pca}{\code{pagoo::PgR6M$gg_pca()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="gg_pie">}\href{../../pagoo/html/PgR6M.html#method-gg_pie}{\code{pagoo::PgR6M$gg_pie()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="pan_pca">}\href{../../pagoo/html/PgR6M.html#method-pan_pca}{\code{pagoo::PgR6M$pan_pca()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="pg_power_law_fit">}\href{../../pagoo/html/PgR6M.html#method-pg_power_law_fit}{\code{pagoo::PgR6M$pg_power_law_fit()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="rarefact">}\href{../../pagoo/html/PgR6M.html#method-rarefact}{\code{pagoo::PgR6M$rarefact()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6M" data-id="runShinyApp">}\href{../../pagoo/html/PgR6M.html#method-runShinyApp}{\code{pagoo::PgR6M$runShinyApp()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="pagoo" data-topic="PgR6MS" data-id="core_seqs_4_phylo">}\href{../../pagoo/html/PgR6MS.html#method-core_seqs_4_phylo}{\code{pagoo::PgR6MS$core_seqs_4_phylo()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PewitR6$new(
  data,
  org_meta,
  cluster_meta,
  core_level = 95,
  sep = "__",
  sequences,
  verbose = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gene_context"></a>}}
\if{latex}{\out{\hypertarget{method-gene_context}{}}}
\subsection{Method \code{gene_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PewitR6$gene_context(cluster, upstream = 2, downstream = 2)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gg_gene_context"></a>}}
\if{latex}{\out{\hypertarget{method-gg_gene_context}{}}}
\subsection{Method \code{gg_gene_context()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PewitR6$gg_gene_context(
  cluster,
  upstream = 2,
  downstream = 2,
  fill = "cluster",
  orient = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PewitR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
