% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome.R
\name{plot.pangenome}
\alias{plot.pangenome}
\title{Plot a \code{pangenome} object.}
\usage{
\method{plot}{pangenome}(x, hclustfun = hclust, distfun = dist,
  edgePar = list(lwd = 1.5), margins = c(5, 7), col = c("white",
  "#E64B35FF"), raster = TRUE, labRow = NULL, cexRow = 0.5,
  colSideColors = c("#4DBBD5FF", "#00A087FF", "#3C5488FF"),
  colSideLabels = c("Singletones", "Accessory", "Soft-Core Genome"),
  cexKey = 1, ...)
}
\arguments{
\item{x}{\code{pangenome} object.}

\item{hclustfun}{Function used to compute the hierarchical clustering.}

\item{distfun}{Function used to compute the distance (dissimilarity).}

\item{edgePar}{A list of parameters to be passed to plot.dendrogram().}

\item{margins}{A \code{numeric} vector of length 2 containing the margins
for column and row names, repectively.}

\item{col}{A \code{character} vector with the colors of the presence/absence
plot. The first color indicate absence, and the second indicates presence.}

\item{raster}{\code{logical} See \code{useRaster} parameter in \code{?image}.}

\item{labRow}{A \code{character} vector of length equal to the number of
organisms in the pangenome object. By default it uses the original file
names without the extension (.gff).}

\item{cexRow}{\code{numeric} The size of row labels.}

\item{colSideColors}{A \code{character} vector of length 3 specifying the
colors to use on the column bar. The colors correspond, in order, to
singletons, accessory and soft-core clusters.}

\item{colSideLabels}{A \code{character} vector of length 3 to specify the
labels to use to indicate Singletons, Accessory genes, and Soft-core genes.}

\item{cexKey}{\code{numeric} The size of the key labels.}

\item{...}{Unused}
}
\description{
Generic function to plot a \code{pangenome} object.
}
\details{
A heatmap-like image is plotted attached with a dendrogram. The
grid shows the presence or absence of a gene (column), for each
organism (rows). The columns are ordered so the core genes appear at the
left, the accessory genes at the middle, and the singletones at the right
of the grid.

The left dendrogram is computed based on a dissimilarity measure between
the vectors of presence/absence of 2 genomes at a time.

The top bar indicates the genes (columns) assigned as a part of the soft-
core genome, the accessory genome, and the singletones.
}
\author{
Ignacio Ferres
}
