% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{plotRarefaction}
\alias{plotRarefaction}
\title{Plot Pangenome Rarefaction Curves}
\usage{
plotRarefaction(x, nsamp = 10, plot = TRUE, xlab = "# of genomes",
  ylab = "# of genes", y.mar = c(0.85, 1.15), las = 0,
  pg.col = "#E64B35FF", cg.col = "#4DBBD5FF", pt.pch = 20,
  pt.cex = 1, pt.alpha = 0.4, li.lwd = 3, shadow = TRUE,
  shadow.quant = c(2, 4), shadow.col = "grey", leg = TRUE,
  legend = c("Pangenome", "Coregenome"), leg.pos = "topleft",
  leg.cex = par("cex"), leg.pt.cex = par("cex"), leg.pch = 22,
  leg.bty = "n", ...)
}
\arguments{
\item{x}{An object of class \code{pangenome}.}

\item{nsamp}{\code{integer} The number of genomes to sample on each stage.}

\item{plot}{\code{logical} If \code{TRUE}, then a plot is produced. If not,
just the result matrices are returned invisibly.}

\item{xlab}{\code{character} X axis label.}

\item{ylab}{\code{character} Y axis label.}

\item{y.mar}{A \code{vector} of length 2, indicating a proportion of the
minimum and maximum values to take as bottom and top margins, respectively.}

\item{las}{\code{numeric} Style of axis labels. See \code{?par}.}

\item{pg.col}{Pangenome points and curve color.}

\item{cg.col}{Coregenome points and curve color.}

\item{pt.pch}{An \code{integer} specifying a symbol. See \code{?par}.}

\item{pt.cex}{\code{numeric} The point size.}

\item{pt.alpha}{A factor modifying the opacity alpha of the points.}

\item{li.lwd}{\code{integer} The line width.}

\item{shadow}{\code{logical} Plot a ribbon showing a function of points
quantiles (quantiles to plot specifyied by \code{shadow.quant}).}

\item{shadow.quant}{A \code{vector} of length 2 in {0, 1, 2, 3, 4, 5},
specifying the quantiles to plot.}

\item{shadow.col}{The ribbon color.}

\item{leg}{\code{logical} Plot a legend?}

\item{legend}{A \code{vector} of length 2 with the legend names to plot.}

\item{leg.pos}{The position of the legend.}

\item{leg.cex}{\code{numeric} The legend size.}

\item{leg.pt.cex}{\code{numeric} The legend point size.}

\item{leg.pch}{An \code{integer} specifying the legend symbols.}

\item{leg.bty}{The legend box type. See \code{?legend}.}

\item{...}{Arguments to be passed to \code{plot()}}
}
\value{
A \code{list} of 2 \code{nsamp}*# of organisms \code{matrix} is
returned.
}
\description{
Plot tipical pangenome and coregenome curves.
}
\details{
Both the number of shared genes and the total number of genes as
a function of the number of organisms sequencially added are plotted. For
each new genome added, a sample of \code{nsamp} (\code{default} 10) genomes
are evaluated with no replace.

A scatter plot is drawn.
Future versions will allow more customization.

Also 2 matrices are invisibly returned, the first for the core rarefaction
curve, and the second for the pangenome rarefaction curve. Each cell is the
count of either core or pan genes for the ith sample (rows) of jth organism
(columns) added.
}
