% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{processPfam_A_Dat}
\alias{processPfam_A_Dat}
\title{Process Pfam-A.dat file}
\usage{
processPfam_A_Dat(datPfam, n_threads)
}
\arguments{
\item{datPfam}{\code{character}. The path to Pfam-A.dat file.}

\item{n_threads}{\code{integer}. The number of threads to use.}
}
\value{
A \code{data.frame} with information of each Pfam entry.
}
\description{
Process Pfam-A.dat file.
}
\author{
Ignacio Ferres
}
