% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pangenome.R
\name{summary.pangenome}
\alias{summary.pangenome}
\title{Summarizing pangenome object}
\usage{
\method{summary}{pangenome}(object, ...)
}
\arguments{
\item{object}{An object of class \code{pangenome}}

\item{...}{Unused}
}
\value{
A \code{matrix} with softcore, accessory and singletons numbers for
each genome.
}
\description{
Summary method for class \code{pangenome}
}
\author{
Ignacio Ferres
}
